requirements.txt
apache-airflow
pandas
scikit-learn
numpy
joblib
python-dotenv
